Splitting up `feedItems.types.ts` even more
