+++ packages/shared/src/parsers/accounts.parser.ts
@@ -6,7 +6,8 @@ import type {Account} from '@shared/types/accounts.types';
 import type {AccountId} from '@shared/types/ids.types';
 import type {Result} from '@shared/types/results.types';
 
-import {AccountIdSchema, AccountSchema} from '@shared/schemas/accounts.schema';
+import {AccountSchema} from '@shared/schemas/accounts.schema';
+import {AccountIdSchema} from '@shared/schemas/ids.schema';
 import {fromStorageAccount} from '@shared/storage/accounts.storage';
 
 /**
