+++ packages/shared/src/parsers/feedSubscriptions.parser.ts
@@ -10,10 +10,10 @@ import type {FeedSubscriptionId} from '@shared/types/ids.types';
 import type {Result} from '@shared/types/results.types';
 
 import {
-  FeedSubscriptionIdSchema,
   FeedSubscriptionLifecycleSchema,
   FeedSubscriptionSchema,
 } from '@shared/schemas/feedSubscriptions.schema';
+import {FeedSubscriptionIdSchema} from '@shared/schemas/ids.schema';
 import {fromStorageFeedSubscription} from '@shared/storage/feedSubscriptions.storage';
 
 /**
