+++ packages/shared/src/schemas/accounts.schema.ts
@@ -1,9 +1,8 @@
 import {z} from 'zod/v4';
 
 import {EmailAddressSchema} from '@shared/schemas/emails.schema';
+import {AccountIdSchema} from '@shared/schemas/ids.schema';
 import {BaseStoreItemSchema} from '@shared/schemas/utils.schema';
-
-export const AccountIdSchema = z.string().min(1).max(128);
 
 export const AccountSchema = BaseStoreItemSchema.extend({
   accountId: AccountIdSchema,
