+++ packages/shared/src/schemas/feedItems.schema.ts
@@ -1,9 +1,8 @@
 import {z} from 'zod/v4';
 
 import {FeedItemContentType} from '@shared/types/feedItemContent.types';
-import {FeedItemImportStatus, TriageStatus} from '@shared/types/feedItems.types';
+import {TriageStatus} from '@shared/types/feedItems.types';
 
-import {AccountIdSchema} from '@shared/schemas/accounts.schema';
 import {
   ArticleFeedItemContentSchema,
   IntervalFeedItemContentSchema,
@@ -13,58 +12,11 @@ import {
   XkcdFeedItemContentSchema,
   YouTubeFeedItemContentSchema,
 } from '@shared/schemas/feedItemContent.schema';
+import {FeedItemImportStateSchema} from '@shared/schemas/feedItemImportStates.schema';
 import {FeedSchema} from '@shared/schemas/feeds.schema';
-import {FirestoreTimestampSchema} from '@shared/schemas/firebase.schema';
+import {AccountIdSchema, FeedItemIdSchema} from '@shared/schemas/ids.schema';
 import {BaseStoreItemSchema} from '@shared/schemas/utils.schema';
 
-export const FeedItemIdSchema = z.uuid();
-
-//////////////////////////////
-//  FEED ITEM IMPORT STATE  //
-//////////////////////////////
-const NewFeedItemImportStateSchema = z.object({
-  status: z.literal(FeedItemImportStatus.New),
-  shouldFetch: z.literal(true),
-  lastSuccessfulImportTime: z.null(),
-  lastImportRequestedTime: FirestoreTimestampSchema.or(z.date()),
-});
-
-const ProcessingFeedItemImportStateSchema = z.object({
-  status: z.literal(FeedItemImportStatus.Processing),
-  shouldFetch: z.literal(false),
-  importStartedTime: FirestoreTimestampSchema.or(z.date()),
-  lastSuccessfulImportTime: FirestoreTimestampSchema.or(z.date()).or(z.null()),
-  lastImportRequestedTime: FirestoreTimestampSchema.or(z.date()),
-});
-
-const FailedFeedItemImportStateSchema = z.object({
-  status: z.literal(FeedItemImportStatus.Failed),
-  shouldFetch: z.boolean(),
-  errorMessage: z.string(),
-  importFailedTime: FirestoreTimestampSchema.or(z.date()),
-  lastSuccessfulImportTime: FirestoreTimestampSchema.or(z.date()).or(z.null()),
-  lastImportRequestedTime: FirestoreTimestampSchema.or(z.date()),
-});
-
-const CompletedFeedItemImportStateSchema = z.object({
-  status: z.literal(FeedItemImportStatus.Completed),
-  shouldFetch: z.boolean(),
-  lastSuccessfulImportTime: FirestoreTimestampSchema.or(z.date()),
-  lastImportRequestedTime: FirestoreTimestampSchema.or(z.date()),
-});
-
-const FeedItemImportStateSchema = z.discriminatedUnion('status', [
-  NewFeedItemImportStateSchema,
-  ProcessingFeedItemImportStateSchema,
-  FailedFeedItemImportStateSchema,
-  CompletedFeedItemImportStateSchema,
-]);
-
-export type FeedItemImportStateFromStorage = z.infer<typeof FeedItemImportStateSchema>;
-
-/////////////////
-//  FEED ITEM  //
-/////////////////
 const BaseFeedItemSchema = BaseStoreItemSchema.extend({
   origin: FeedSchema,
   feedItemId: FeedItemIdSchema,
