+++ packages/shared/src/schemas/tags.schema.ts
@@ -2,9 +2,8 @@ import {z} from 'zod/v4';
 
 import {SystemTagId} from '@shared/types/tags.types';
 
+import {UserTagIdSchema} from '@shared/schemas/ids.schema';
 import {BaseStoreItemSchema} from '@shared/schemas/utils.schema';
-
-export const UserTagIdSchema = z.uuid();
 
 export const SystemTagIdSchema = z.enum(SystemTagId);
